+++ small_text/query_strategies/strategies.py
@@ -1,19 +1,15 @@
 from abc import ABC, abstractmethod
-from typing import Union, TYPE_CHECKING
+from typing import Union
 
 import numpy as np
-
-if TYPE_CHECKING:
-    import numpy.typing as npt
-
-    from small_text.classifiers import Classifier
-    from small_text.data import Dataset
-    from scipy.sparse import csr_matrix
-
+import numpy.typing as npt
+from scipy.sparse import csr_matrix
 from scipy.stats import entropy
 from sklearn.preprocessing import normalize
 
 from small_text.base import check_optional_dependency
+from small_text.classifiers import Classifier
+from small_text.data import Dataset
 from small_text.query_strategies.exceptions import EmptyPoolException, PoolExhaustedException
 from small_text.utils.context import build_pbar_context