+++ small_text/utils/clustering.py
@@ -1,12 +1,25 @@
 import warnings
 import numpy as np
 
+from packaging.version import parse, Version
+from typing import TYPE_CHECKING, Union
+
 from numpy.random import choice
 from sklearn.cluster import kmeans_plusplus
 
+# if TYPE_CHECKING:
+import numpy.typing as npt
+from numpy.random import RandomState
 
-def init_kmeans_plusplus_safe(x, n_clusters, *, x_squared_norms=None, random_state=None,
-                              n_local_trials=None, suppress_warning=False):
+
+def init_kmeans_plusplus_safe(x: npt.NDArray[np.double],
+                              n_clusters: int,
+                              *,
+                              weights: Union[npt.NDArray[np.double], None] = None,
+                              x_squared_norms: Union[npt.NDArray[np.double], None] = None,
+                              random_state: Union[int, RandomState, None] = None,
+                              n_local_trials: Union[int, None] = None,
+                              suppress_warning: bool = False):
     """
     Calls scikit-learn's k-means++ initialization with a fallback that prevents duplicate
     cluster centers. If duplicate indices are encountered, they are replaced with randomly sampled
@@ -14,17 +27,23 @@ def init_kmeans_plusplus_safe(x, n_clusters, *, x_squared_norms=None, random_sta
 
     Parameters
     ----------
-    x : numpy.ndarray
+    x : numpy.ndarray[double]
         The input data to be clustered.
     n_clusters : int
         The number of cluster centers to find.
-    x_squared_norms : numpy.ndarray
+    weights : numpy.ndarray[double], default=None
+
+        Sample weights or None. Using this requires scikit-learn >= 1.3.0.
+
+        .. versionadded:: 2.0.0
+
+    x_squared_norms : numpy.ndarray[double], default=None
         List of squared norms or None.
-    random_state : int or RandomState instance
+    random_state : int or RandomState instance or None, default=None
         A random state or None.
-    n_local_trials : int
+    n_local_trials : int, default=None
         Number of local trials.
-    suppress_warning : bool
+    suppress_warning : bool, default=None
         Suppresses the warning given on duplicate indices if True.
 
     Returns
@@ -39,12 +58,20 @@ def init_kmeans_plusplus_safe(x, n_clusters, *, x_squared_norms=None, random_sta
     sklearn.cluster import kmeans_plusplus :
         The actual k-means++ initialization provided by scikit-learn.
     """
+    kmeans_plusplus_kwargs = {}
+    if weights is not None:
+        from sklearn import __version__
+        if parse(__version__) < Version('1.3.0'):
+            raise ValueError('scikit-learn>=1.3.0 is required to use the "weights" argument.')
+
+        kmeans_plusplus_kwargs['sample_weight'] = weights
 
     centers, indices = kmeans_plusplus(x,
                                        n_clusters,
                                        x_squared_norms=x_squared_norms,
                                        random_state=random_state,
-                                       n_local_trials=n_local_trials)
+                                       n_local_trials=n_local_trials,
+                                       **kmeans_plusplus_kwargs)
 
     unique_indices, counts = np.unique(indices, return_counts=True)
 
@@ -57,8 +84,8 @@ def init_kmeans_plusplus_safe(x, n_clusters, *, x_squared_norms=None, random_sta
 
         centers = np.delete(centers, np.arange(unique_indices.shape[0])[counts > 1], axis=0)
 
-        fill_indices = choice(remaining_indices, size=n_clusters-unique_indices.shape[0],
-                              replace=False)
+        fill_indices = choice(remaining_indices, size=n_clusters-unique_indices.shape[0], replace=False)
+
         indices = np.hstack((unique_indices, fill_indices))
         centers = np.vstack((centers, x[fill_indices]))