+++ CHANGELOG.md
@@ -32,6 +32,9 @@ On the other hand, this also allowed us to deal with further issues that contain
 - General
   - Moved `split_data()` method from `small_text.data.datasets` to `small_text.data.splits`.
 
+- Utils
+  - `init_kmeans_plusplus_safe()` now supports weighted kmeans++ initialization for `scikit-learn>=1.3.0`. 
+
 ### Removed
 
 - Removed `default_tensor_type()` method.